fix(review): address PR review comments
- Extract getDsfrModal to ~/modules/shared to eliminate duplication across Step2Upload and JointEvaluationForm
- Fix NewTabNotice import to use ~/modules/layout barrel instead of internal path
- Replace custom SCSS classes with DSFR components: fr-highlight for deadline block, fr-callout/fr-callout--blue-france for info boxes
- Delete JointEvaluationForm.module.scss (no longer needed)
