--- packages/app/src/modules/declaration-remuneration/steps/jointEvaluation/JointEvaluationForm.module.scss
@@ -1,12 +1,0 @@
-.deadline {
-	padding-left: 2rem;
-	border-left: 4px solid var(--border-plain-blue-france);
-}
-
-.infoBox {
-	padding: 1.875rem;
-}
-
-.infoBoxBlue {
-	background: var(--background-alt-blue-france);
-}
