+++ packages/app/src/modules/shared/index.ts
@@ -1,1 +1,2 @@
+export { getDsfrModal } from "./getDsfrModal";
 export { PdfFileUpload } from "./PdfFileUpload";
